Fix a problem with the definition of union semun which broke System V semaphores on systems where sizeof(int) != sizeof(long).
PR:               12072
Submitted by:	  <winterling@de.ibm.com>
Reviewed by:	  Jeff Trawick, Justin Erenkrantz


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@96927 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_main.c
@@ -749,9 +749,8 @@ accept_mutex_methods_s accept_mutex_pthread_s = {
 #include <sys/sem.h>
 
 #ifdef NEED_UNION_SEMUN
-/* it makes no sense, but this isn't defined on solaris */
 union semun {
-    long val;
+    int val;
     struct semid_ds *buf;
     ushort *array;
 };
